% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirementDash.R
\docType{package}
\name{retirementDash}
\alias{retirementDash}
\title{retirementDash:  Finding retirement locations in the US.}
\description{
The retirementDash package can provide and narrow prospective retirement counties by important
demographic, environmental, housing affordability and other characteristics.
}
\seealso{
Useful links:
\itemize{
\item \url{https://robwiederstein.github.io/retirementDash/}
\item Report bugs at \url{https://github.com/RobWiederstein/retirementDash/issues}
}
}
\author{
Rob Wiederstein
}
