% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retirementDash.R
\docType{package}
\name{retirementDash}
\alias{retirementDash}
\title{retirementDash:  Finding retirement locations in the US.}
\description{
The retirementDash package can provide and narrow prospective counties by important
demographic, environmental and other characteristics.
}
\seealso{
Useful links:
\itemize{
\item https://github.com/robwiederstein/retirementDash/www.github.repo.com
\item Report bugs at https://github.com/robwiederstein/retirementDash/issues
}
}
\author{
Rob Wiederstein
}
